requirements.txt
fastapi
pydantic
motor
pymongo
redis
kafka-python
uvicorn
